+++ packages/salesforcedx-vscode-lwc/test/vscode-integration/commands/forceLightningLwcPreview.test.ts
@@ -13,12 +13,13 @@ import * as vscode from 'vscode';
 import { DEV_SERVER_PREVIEW_ROUTE } from '../../../src/commands/commandConstants';
 import * as commandUtils from '../../../src/commands/commandUtils';
 import { forceLightningLwcPreview } from '../../../src/commands/forceLightningLwcPreview';
+import { nls } from '../../../src/messages';
 import { DevServerService } from '../../../src/service/devServerService';
 
 const sfdxCoreExports = vscode.extensions.getExtension(
   'salesforce.salesforcedx-vscode-core'
 )!.exports;
-const { SfdxCommandlet } = sfdxCoreExports;
+const { SfdxCommandlet, notificationService } = sfdxCoreExports;
 
 describe('forceLightningLwcPreview', () => {
   let sandbox: SinonSandbox;
@@ -26,6 +27,7 @@ describe('forceLightningLwcPreview', () => {
   let openBrowserStub: SinonStub;
   let existsSyncStub: sinon.SinonStub;
   let lstatSyncStub: sinon.SinonStub;
+  let showErrorMessageStub: sinon.SinonStub;
 
   beforeEach(() => {
     sandbox = sinon.sandbox.create();
@@ -34,6 +36,10 @@ describe('forceLightningLwcPreview', () => {
     openBrowserStub = sandbox.stub(commandUtils, 'openBrowser');
     existsSyncStub = sandbox.stub(fs, 'existsSync');
     lstatSyncStub = sandbox.stub(fs, 'lstatSync');
+    showErrorMessageStub = sandbox.stub(
+      notificationService,
+      'showErrorMessage'
+    );
   });
 
   afterEach(() => {
@@ -126,5 +132,79 @@ describe('forceLightningLwcPreview', () => {
     await forceLightningLwcPreview(sourceUri);
 
     sinon.assert.calledOnce(commandletStub);
+  });
+
+  it('shows an error when source path is not recognized as an lwc module file', async () => {
+    devServiceStub.isServerHandlerRegistered.returns(true);
+
+    const testPath = path.join('foo');
+    const sourceUri = { path: testPath } as vscode.Uri;
+
+    existsSyncStub.returns(true);
+    lstatSyncStub.returns({
+      isDirectory() {
+        return false;
+      }
+    });
+
+    await forceLightningLwcPreview(sourceUri);
+
+    sinon.assert.calledWith(
+      showErrorMessageStub,
+      sinon.match(
+        nls.localize(`force_lightning_lwc_preview_unsupported`, 'foo')
+      )
+    );
+  });
+
+  it('shows an error when source path does not exist', async () => {
+    devServiceStub.isServerHandlerRegistered.returns(true);
+
+    const testPath = path.join('foo');
+    const sourceUri = { path: testPath } as vscode.Uri;
+
+    existsSyncStub.returns(false);
+
+    await forceLightningLwcPreview(sourceUri);
+
+    sinon.assert.calledWith(
+      showErrorMessageStub,
+      sinon.match(
+        nls.localize(`force_lightning_lwc_preview_file_nonexist`, 'foo')
+      )
+    );
+  });
+
+  it('shows an error message when open browser throws an error', async () => {
+    devServiceStub.isServerHandlerRegistered.returns(true);
+
+    const testPath = path.join(
+      'dev',
+      'project',
+      'force-app',
+      'main',
+      'default',
+      'lwc',
+      'foo'
+    );
+    const sourceUri = { path: testPath } as vscode.Uri;
+
+    existsSyncStub.returns(true);
+    lstatSyncStub.returns({
+      isDirectory() {
+        return true;
+      }
+    });
+
+    openBrowserStub.throws('test error');
+
+    await forceLightningLwcPreview(sourceUri);
+
+    const commandName = nls.localize(`force_lightning_lwc_preview_text`);
+    sinon.assert.calledTwice(showErrorMessageStub);
+    sinon.assert.calledWith(
+      showErrorMessageStub,
+      sinon.match(nls.localize('command_failure', commandName))
+    );
   });
 });